pyproject.toml
[project]
name = "arquitetura-dw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-apache-kafka==1.9.0",
    "apache-airflow-providers-common-sql==1.27.2",
    "apache-airflow-providers-postgres==6.2.0",
    "astronomer-cosmos==1.10.1",
    "duckdb>=1.3.1",
    "pandas>=2.3.0",
]
